3955 Inline workspace_id on data_table and remove workspace_data_table
