--- server/libs/automation/automation-data-table/automation-data-table-api/src/main/java/com/bytechef/automation/data/table/configuration/domain/WorkspaceDataTable.java
@@ -1,137 +1,0 @@
-/*
- * Copyright 2025 ByteChef
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.bytechef.automation.data.table.configuration.domain;
-
-import java.time.Instant;
-import java.util.Objects;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.annotation.Version;
-import org.springframework.data.relational.core.mapping.Column;
-import org.springframework.data.relational.core.mapping.Table;
-
-/**
- * @author Ivica Cardic
- */
-@Table("workspace_data_table")
-public class WorkspaceDataTable {
-
-    @Column("data_table_id")
-    private Long dataTableId;
-
-    @CreatedBy
-    @Column("created_by")
-    private String createdBy;
-
-    @Column("created_date")
-    @CreatedDate
-    private Instant createdDate;
-
-    @Id
-    private Long id;
-
-    @Column("last_modified_by")
-    @LastModifiedBy
-    private String lastModifiedBy;
-
-    @Column("last_modified_date")
-    @LastModifiedDate
-    private Instant lastModifiedDate;
-
-    @Column("workspace_id")
-    private Long workspaceId;
-
-    @Version
-    private int version;
-
-    private WorkspaceDataTable() {
-    }
-
-    public WorkspaceDataTable(Long dataTableId, Long workspaceId) {
-        this.dataTableId = dataTableId;
-        this.workspaceId = workspaceId;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-
-        WorkspaceDataTable that = (WorkspaceDataTable) o;
-
-        return Objects.equals(id, that.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-
-    public Long getDataTableId() {
-        return dataTableId;
-    }
-
-    public String getCreatedBy() {
-        return createdBy;
-    }
-
-    public Instant getCreatedDate() {
-        return createdDate;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getLastModifiedBy() {
-        return lastModifiedBy;
-    }
-
-    public Instant getLastModifiedDate() {
-        return lastModifiedDate;
-    }
-
-    public Long getWorkspaceId() {
-        return workspaceId;
-    }
-
-    public int getVersion() {
-        return version;
-    }
-
-    @Override
-    public String toString() {
-        return "WorkspaceDataTable{" +
-            "id=" + id +
-            ", dataTableId=" + dataTableId +
-            ", workspaceId=" + workspaceId +
-            ", createdBy='" + createdBy + '\'' +
-            ", createdDate=" + createdDate +
-            ", lastModifiedBy='" + lastModifiedBy + '\'' +
-            ", lastModifiedDate=" + lastModifiedDate +
-            ", version=" + version +
-            '}';
-    }
-}
